% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfoundingExplorer-package.R
\docType{package}
\name{ConfoundingExplorer-package}
\alias{ConfoundingExplorer-package}
\title{ConfoundingExplorer}
\description{
ConfoundingExplorer is an R package for exploring the impact of confounding
between a condition of interest and a batch variable.
}
\author{
\strong{Maintainer}: Charlotte Soneson \email{charlottesoneson@gmail.com} (\href{https://orcid.org/0000-0003-3833-2169}{ORCID})

}
\keyword{internal}
