% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfoundingExplorer-pkg.R
\docType{package}
\name{ConfoundingExplorer-pkg}
\alias{ConfoundingExplorer-pkg}
\title{ConfoundingExplorer}
\description{
ConfoundingExplorer is an R package for exploring the impact of confounding
between a condition of interest and a batch variable.
}
