% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{.generateData}
\alias{.generateData}
\title{Generate data}
\usage{
.generateData(
  nb1c1,
  nb1c2,
  nb2c1,
  nb2c2,
  fraccond,
  fracbatch,
  fracunknown,
  condeffect,
  batcheffect,
  unknowneffect,
  unknowntype,
  analysisapproach,
  nvar,
  seed
)
}
\arguments{
\item{nb1c1, nb1c2, nb2c1, nb2c2}{Number of samples in the respective groups
(batch 1/2, condition 1/2)}

\item{fraccond, fracbatch}{Fraction of variables affected by condition and
batch, respectively}

\item{condeffect, batcheffect}{Condition and batch effect size, respectively}

\item{analysisapproach}{Type of statistical model to fit}

\item{nvar}{Number of variables}

\item{seed}{Random seed}
}
\value{
A list with three elements: the simulated data matrix (m), the
  results of the statistical test (res) and the sample annotation table
  (annot).
}
\description{
Generate data
}
\author{
Charlotte Soneson
}
\keyword{internal}
