% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfoundingExplorer.R
\name{ConfoundingExplorer}
\alias{ConfoundingExplorer}
\title{Confounding explorer}
\usage{
ConfoundingExplorer()
}
\value{
A shinyApp object
}
\description{
Confounding explorer
}
\examples{
if (interactive()) {
  ConfoundingExplorer()
}

}
\author{
Charlotte Soneson
}
